requirements.txt
numpy>=1.20.0
boto3
cloud-volume>=0.47.0
click>=8.1 
tqdm>=4.36.1
connected-components-3d
scikit-image>=0.14.2
scikit-learn
h5py
neuroglancer>=2.28
tinybrain
tifffile
pandas>=1.3.5
zmesh
fastremap
matplotlib
kimimaro
pyspng-seunglab~=1.0.0
pyparsing>=2.4.2
urllib3>=1.25.11
docutils>=0.10
pynrrd
traitlets
